Crate srml_nicks

Nicks Module

nicks::Trait

Call

Overview

Nicks is a trivial module for keeping track of account names on-chain. It makes no effort to create a name hierarchy, be a DNS replacement or provide reverse lookups.

Interface

Dispatchable Functions

set_name - Set the associated name of an account; a small deposit is reserved if not already taken.

clear_name - Remove an account's associated name; the deposit is returned.

kill_name - Forcibly remove the associated name; the deposit is lost.

Structs

Module

Enums

Call

RawEvent
Events for this module.

Traits

Trait

Type Definitions

Event
RawEvent specialized for the configuration Trait